backend/scraper/requirements.txt
arxiv==2.1.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.3
google-generativeai>=0.3.0
groq>=0.4.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
